requirements.txt
pyjwt == 1.7.1
cryptography == 2.8
requests == 2.22.0
